+++ src/hooks/useAuth.tsx
@@ -1,0 +1,7 @@
+import { useAppSelector } from '@/redux/hooks';
+import { selectIsAuth } from '@/redux/slices/user-slice';
+
+export const useAuth = () => {
+  const isAuth = useAppSelector(selectIsAuth);
+  return isAuth;
+};
